refactor: changed package name and structure for better interpretability
