+++ Uni_life_tracker/app/src/test/java/com/generic/plannr/tests/usecase/AddEventTest.java
@@ -1,4 +1,4 @@
-package com.generic.plannr.usecasetests;
+package com.generic.plannr.tests.usecase;
 
 import org.junit.*;
