+++ Uni_life_tracker/app/src/test/java/com/generic/plannr/tests/usecase/ExpensesManagerTest.java
@@ -1,4 +1,4 @@
-package com.generic.plannr.usecasetests;
+package com.generic.plannr.tests.usecase;
 
 import com.generic.plannr.Entities.Expense;
 import com.generic.plannr.UseCases.ExpensesManager;
